+++ src/common/base.js
@@ -1,5 +1,4 @@
 const fs = require("fs-extra");
-
 const _ = require("lodash/fp");
 const glob = require("glob");
 const resolve = require("path").resolve;
@@ -28,7 +27,14 @@ obj.fromGlobsToFilesArray = globArray => {
   return globArray.reduce((acc, path) => acc.concat(glob.sync(path)), []);
 };
 
-obj.exportFile = async (code, file, env, ease = false, done) => {
+obj.exportFile = async (
+  code,
+  file,
+  env,
+  ease = false,
+  getFunctionName,
+  done
+) => {
   const obj = utils.getDBObjectFromPath(file);
   const connCfg = db.getConfiguration(env, obj.owner);
   // Owner can change to default user
@@ -41,7 +47,7 @@ obj.exportFile = async (code, file, env, ease = false, done) => {
     // try {
     if (!ease || (await db.isDifferentDdlTime(conn, obj))) {
       // Get Db object code as string
-      let lob = await db.getObjectDdl(conn, obj);
+      let lob = await db.getObjectDdl(conn, getFunctionName, obj);
       lob = _.pipe(
         // Remove whitespaces
         _.trim,
@@ -95,7 +101,7 @@ function getLineAndPosition(code, offset) {
   return { line, position };
 }
 
-obj.compileFile = async (code, file, env, force = false) => {
+obj.compileFile = async (code, file, env, force = false, warnings) => {
   const obj = utils.getDBObjectFromPath(file);
   const connCfg = db.getConfiguration(env, obj.owner);
   obj.owner = connCfg.user.toUpperCase();
@@ -116,7 +122,7 @@ obj.compileFile = async (code, file, env, force = false) => {
       errors = db.getErrorObjectChanged();
     } else {
       // Otherwise compile object to Db with warning scope
-      result = await db.compile(conn, code.toString());
+      result = await db.compile(conn, code.toString(), warnings);
       // Mark object as exported as we have the latest version
       if (!force) await db.syncDdlTime(conn, obj);
       // Getting errors for this object from Db
@@ -244,6 +250,29 @@ obj.resolveObjectInfo = async (env, { name }) => {
     conn && conn.close();
   }
   return result;
+};
+
+obj.getGenerator = async ({ func, file, env, object }) => {
+  const obj = utils.getDBObjectFromPath(file);
+  const connCfg = db.getConfiguration(env, obj.owner);
+  obj.owner = connCfg.user.toUpperCase();
+
+  let result = {};
+  let conn;
+  try {
+    conn = await db.getConnection(connCfg);
+    result = await db.getGeneratorFunction(conn, func, obj, object);
+  } catch (error) {
+    throw error;
+  } finally {
+    conn && conn.close();
+  }
+  return {
+    obj,
+    file,
+    env,
+    result
+  };
 };
 
 module.exports = obj;