+++ src/gulpfile.js
@@ -19,12 +19,12 @@ const inquirer = require("inquirer");
 const multiDest = require("gulp-multi-dest");
 const Table = require("cli-table");
 
-const utils = require("./config/utils/utility");
-const git = require("./config/utils/git");
-const base = require("./config/utils/base");
-const db = require("./config/utils/db");
+const utils = require("./common/utility");
+const git = require("./common/git");
+const base = require("./common/base");
+const db = require("./common/db");
 
-let config = utils.config;
+let config = new utils.Config();
 
 const generateChangeLog = function(paths) {
   // Create Db objects from paths array
@@ -82,7 +82,7 @@ PROMPT ***********************************************************
   done(null, prompt + code + ending);
 };
 
-const packageSrcFromFile = () => {
+const packageSrcFromFile = ({ env = argv.env }) => {
   const deployPrepend = `
 SPOOL deploy.log
 SET FEEDBACK ON
@@ -95,21 +95,25 @@ COMMIT;
 SPOOL OFF
 `;
 
-  const deployFile = path.basename(config.get("package.output"));
-  const deployDir = path.dirname(config.get("package.output"));
+  const outputFileName = path.basename(
+    config.get({ field: "package.output", env })
+  );
+  const outputDirectory = path.dirname(
+    config.get({ field: "package.output", env })
+  );
 
-  const src = config.get("package.input");
-  const encoding = config.get("package.encoding");
+  const src = config.get({ field: "package.input", env });
+  const encoding = config.get({ field: "package.encoding", env });
 
   const templateObject = {
-    config: config.get(),
+    config: config.get({ env }),
     data: {
       // Dates in format YYYY-MM-DD
       now: new Date().toISOString().substring(0, 10)
     }
   };
 
-  const templating = config.get("package.templating");
+  const templating = config.get({ field: "package.templating", env });
 
   return (
     gulp
@@ -118,12 +122,18 @@ SPOOL OFF
       .pipe(templating ? template(templateObject) : gutil.noop())
       // Adds object prompt to every file
       .pipe(map(addDBObjectPrompt))
-      .pipe(concat(deployFile))
+      .pipe(concat(outputFileName))
       .pipe(insert.wrap(deployPrepend, deployAppend))
       .pipe(insert.prepend(timestampHeader))
       .pipe(convertEncoding({ to: encoding }))
-      .pipe(gulp.dest(deployDir))
-      .on("end", () => console.log(chalk.green("Package created.")))
+      .pipe(gulp.dest(outputDirectory))
+      .on("end", () =>
+        console.log(
+          `${chalk.green(
+            "Package created:"
+          )} ${outputDirectory}/${outputFileName}`
+        )
+      )
   );
 };
 
@@ -174,13 +184,15 @@ const cherryPickFromJiraTask = async () => {
   console.log(`Files changed: ${stdout}`);
 };
 
-const makeChangeLog = () => {
-  const file = path.join(__dirname, "/config/templates/changelog*.md");
+const makeChangeLog = ({ env = argv.env }) => {
+  const file = path.join(__dirname, "/templates/changelog*.md");
   // Generate change log from deploy input array
-  const all = base.fromGlobsToFilesArray(config.get("package.input"));
+  const all = base.fromGlobsToFilesArray(
+    config.get({ field: "package.input", env })
+  );
   const content = generateChangeLog(all); // await git.getChangeLog();
   const templateObject = {
-    config: config.get(),
+    config: config.get({ env }),
     data: {
       // Dates in format YYYY-MM-DD
       now: new Date().toISOString().substring(0, 10)
@@ -203,13 +215,17 @@ const exportFilesFromDb = async ({
   ease = argv.ease,
   quiet = argv.quiet
 }) => {
-  const src =
-    file || (changed ? await getOnlyChangedFiles() : config.get("source"));
+  const source = config.get({ field: "source", env });
+  const src = file || (changed ? await getOnlyChangedFiles(source) : source);
+  const getFunctionName = config.get({
+    field: "import.getDdlFunction",
+    env
+  });
 
   const processFile = async (code, file, done) => {
     let res;
     try {
-      res = await base.exportFile(code, file, env, ease, done);
+      res = await base.exportFile(code, file, env, ease, getFunctionName, done);
       if (!quiet && res.exported)
         console.log(
           `${chalk.green("Imported")} <= ${res.obj.owner}@${env} $${file}`
@@ -264,12 +280,12 @@ const printResults = resp => {
   if (errMsg) console.log(`${errMsg}`);
 };
 
-const getOnlyChangedFiles = async () => {
+const getOnlyChangedFiles = async source => {
   const stdout = await git.getChangesNotStaged();
   // Get array of changed files from git
   const changed = base.fromStdoutToFilesArray(stdout);
-  // Get array of files matched by glob pattern source array
-  const all = base.fromGlobsToFilesArray(config.get("source"));
+  // Get array of files matched by source array parameter
+  const all = base.fromGlobsToFilesArray(source);
   // Intersection of both arrays
   const inter = _.intersection(all)(changed);
   return inter;
@@ -281,14 +297,21 @@ const compileFilesToDb = async ({
   changed = argv.changed,
   force = argv.force
 }) => {
-  const src =
-    file || (changed ? await getOnlyChangedFiles() : config.get("source"));
+  const source = config.get({ field: "source", env });
+  const src = file || (changed ? await getOnlyChangedFiles(source) : source);
+  const warnings = config.get({ field: "compile.warnings", env });
 
   const processFile = async (file, done) => {
     let resp;
     try {
       // Compile file and get errors
-      resp = await base.compileFile(file.contents, file.path, env, force);
+      resp = await base.compileFile(
+        file.contents,
+        file.path,
+        env,
+        force,
+        warnings
+      );
       printResults(resp);
       // Stage file if no errors
       if (!force && !resp.errors.hasErrors()) {
@@ -317,7 +340,7 @@ const compileFilesToDb = async ({
 };
 
 const saveLogFile = ({ env = argv.env }) => {
-  const deployDir = path.dirname(config.get("package.output"));
+  const deployDir = path.dirname(config.get({ field: "package.output", env }));
   const addEnvToName = path => {
     path.basename = path.basename + env;
   };
@@ -328,7 +351,7 @@ const saveLogFile = ({ env = argv.env }) => {
 };
 
 const runFileOnDb = async ({ file = argv.file, env = argv.env }) => {
-  const src = file || config.get("package.output");
+  const src = file || config.get({ field: "package.output", env });
 
   // Simple output err colorizer
   const colorize = text =>
@@ -352,7 +375,7 @@ const runFileOnDb = async ({ file = argv.file, env = argv.env }) => {
 const createDbConfigFile = async done => {
   if (!fs.existsSync("./dbconfig.json")) {
     fs.copySync(
-      path.join(__dirname, "/config/templates/dbconfig.json"),
+      path.join(__dirname, "/templates/dbconfig.json"),
       "./dbconfig.json"
     );
     await inquirer.prompt([
@@ -373,23 +396,23 @@ const createProjectFiles = () => {
   const scriptsDirs = db.getUsers().map(v => `./scripts/${v}`);
   gulp
     .src([
-      path.join(__dirname, "/config/templates/scripts/initial*.sql"),
-      path.join(__dirname, "/config/templates/scripts/final*.sql")
+      path.join(__dirname, "/templates/scripts/initial*.sql"),
+      path.join(__dirname, "/templates/scripts/final*.sql")
     ])
     .pipe(multiDest(scriptsDirs));
 
   let src = [];
 
   if (!fs.existsSync("./.gitignore"))
-    src.push(path.join(__dirname, "/config/templates/.gitignore"));
+    src.push(path.join(__dirname, "/templates/.gitignore"));
   if (!fs.existsSync("./test"))
-    src.push(path.join(__dirname, "/config/templates/test/*.test.sql"));
+    src.push(path.join(__dirname, "/templates/test/*.test.sql"));
 
   src.length === 0 && src.push("nonvalidfile");
   return gulp
     .src(src, {
       allowEmpty: true,
-      base: path.join(__dirname, "/config/templates/")
+      base: path.join(__dirname, "/templates/")
     })
     .pipe(gulp.dest("."));
 };
@@ -591,8 +614,8 @@ const mergeLocalAndDbChanges = async ({
   env = argv.env,
   changed = argv.changed
 }) => {
-  const src =
-    file || (changed ? await getOnlyChangedFiles() : config.get("source"));
+  const source = config.get({ field: "source", env });
+  const src = file || (changed ? await getOnlyChangedFiles(source) : source);
 
   if (src.length !== 0) {
     try {
@@ -612,7 +635,7 @@ const compileAndMergeFilesToDb = async ({
   changed = argv.changed,
   force = argv.force
 }) => {
-  force = force || config.get("compile.force");
+  force = force || config.get({ field: "compile.force", env });
   try {
     // Compile and get error results
     const results = await compileFilesToDbAsync({ file, env, changed, force });
@@ -672,6 +695,28 @@ const compileObjectToDb = async ({
   // console.log(object);
   let resp = await base.compileSelection(object, file, env, line);
   printResults(resp);
+};
+
+const generate = async ({
+  env = argv.env,
+  func = argv.func,
+  file = argv.file,
+  object = argv.object,
+  output = argv.output
+}) => {
+  try {
+    const resp = await base.getGenerator({ func, file, env, object });
+    const outputPath = output
+      ? path.resolve(output)
+      : path.resolve(
+          `./scripts/${resp.obj.owner}/file_${new Date().getTime()}.sql`
+        );
+
+    await utils.outputFilePromise(outputPath, resp.result);
+    console.log(`${outputPath} ${chalk.green("created.")}`);
+  } catch (err) {
+    console.error(err.message);
+  }
 };
 
 gulp.task(
@@ -739,3 +784,5 @@ runTest.description = "Simple unit testing.";
 gulp.task("runTest", runTest);
 
 // gulp.task("default", "packageSrc");
+
+gulp.task("generate", generate);